added requirements.txt
Added a requirements.txt file to the src, this is actually just contains all of the modules for Python that I have installed on my machine currently. I am mainly using this to make it easer to download all of the modules on my Raspberry Pi so I can attempt to run my bots on that instead of my main machine. Hopefully someday I will restrict this to only the necessary modules....